% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_background.R
\name{background_ids}
\alias{background_ids}
\title{Generate background data for IOS calculation}
\usage{
background_ids(
  id_exp = id_exp,
  id_out = id_out,
  type = c("default", "advanced")[1]
)
}
\arguments{
\item{id_exp}{ID for the exposure in IEU GWAS database.}

\item{id_out}{ID for the outome in IEU GWAS database.}

\item{type}{Set a default list of IDs. 
*`default`: Returns list of IDs of GWAS summary statistics generated by many different consortia and generated from UKB.
*`advanced`: In addition to the default list, it returns list of IDs of GWAS summary data on protein, metabolite, and eQTL levels, as well as GWAS summary statistics imported from the EBI database.}

\item{exp}{Dataset of the instruments for the exposure, obtained using \code{extract_instruments}}

\item{id_bg}{List obtained using \code{background_ids}}
}
\value{
Data frame
}
\description{
\code{background_ids} Returns list of IDs of available GWAS summary statistics. The data is obtained from [IEU GWAS database](https://gwas.mrcieu.ac.uk/).
}
