% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bg_to_wide}
\alias{bg_to_wide}
\title{Convert bg_dat to wide format}
\usage{
bg_to_wide(bg_dat, value_column = "rsq.outcome")
}
\arguments{
\item{bg_dat}{Output from make_background}

\item{value_column}{Which column to use to fill the values of the matrix}
}
\value{
Matrix
}
\description{
Imputes missing values using missMDA::imputePCA function
}
