% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phewas.R
\name{extract_phewas}
\alias{extract_phewas}
\title{Extract information from IEU GWAS database}
\usage{
extract_phewas(snplist = NULL, id_bg = id_bg, nsnp_per_chunk = 5)
}
\arguments{
\item{snplist}{A list of instrumental genetic variants for the exposure.}

\item{id_bg}{A list of available GWAS summary statistics.}

\item{nsnp_per_chunk}{Number of SNPs to be used per phewas scanning.}
}
\value{
Data frame
}
\description{
Returns traits which will be used in making background dataset. The data is obtained from [IEU GWAS database](https://gwas.mrcieu.ac.uk/).
}
