% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hclust_instruments}
\alias{hclust_instruments}
\title{Cluster instruments using hclust}
\usage{
hclust_instruments(
  bg_dat,
  value_column,
  kmax = min(50, length(unique(bg_dat$SNP)))
)
}
\arguments{
\item{bg_dat}{<what param does>}

\item{value_column}{<what param does>}

\item{kmax}{<what param does>}
}
\value{

}
\description{
Cluster instruments using hclust
}
