% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.ios_IOS.R
\name{ios}
\alias{ios}
\title{Calculate index of suspicion}
\usage{
ios(exp = exp_dat, bg = bg_dat)
}
\arguments{
\item{exp_dat}{Instruments for the exposure, obtained using \code{extract_instruments}}

\item{bg_dat}{Effects for the instruments on a set of variables, used to calculate index of suspicion}
}
\value{
Data.frame
}
\description{
Calculate IOS across based on R squared multiple traits and summarising estimates using mean, sd, iqr, median, 95% value, maximum value
}
\examples{
ios(exp=exp_dat, bg=bg_dat)

}
