% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{kmeans_instruments}
\alias{kmeans_instruments}
\title{Cluster instruments using k-means}
\usage{
kmeans_instruments(bg_dat, value_column, kmax = 15, nstart = 50, iter.max = 15)
}
\arguments{
\item{bg_dat}{<what param does>}

\item{value_column}{<what param does>}

\item{kmax}{<what param does>}

\item{nstart}{<what param does>}

\item{iter.max}{<what param does>}
}
\value{

}
\description{
Cluster instruments using k-means
}
