% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_background.R
\name{make_background}
\alias{make_background}
\title{Generate background data for IOS calculation}
\usage{
make_background(exp = exp_dat, id_bg = id_bg)
}
\arguments{
\item{exp}{Dataset of the instruments for the exposure, obtained using \code{extract_instruments}}

\item{id_bg}{List obtained using \code{background_ids}
make_background(exp = exp_dat, id_bg = id_bg)}
}
\value{
Data frame
}
\description{
Generates harmonised dataset of SNP-exposure and SNP-other traits that are possibly associated with suspicious SNPs. It also calculates R squared of SNP-exposure and/or SNP-other traits.
}
