% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{mr_cluster}
\alias{mr_cluster}
\title{MR in each cluster group}
\usage{
mr_cluster(bg = bg_dat, dat = dat, method = c("kmean", "hclust", "pv", "pam"))
}
\arguments{
\item{bg}{Effects for the instruments on a set of variables, used to calculate index of suspicion, obtained using \code{make_background()}.}

\item{dat}{Harmonised dataset of the exposure and the outcome, obtained using \code{TwoSampleMR::harmonise_data()}.}

\item{method}{Choose the method to generate cluster. 
\itemize{
  \item kmean See ?\code{kmeans_instruments}
  \item hclust See ?\code{hclust_instruments}
  \item pv See ?\code{pvclust_instruments}
  \item pam See ?\code{pam_instruments}
}}
}
\value{

}
\description{
MR in each cluster group
}
