% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{mr_cluster_estimate}
\alias{mr_cluster_estimate}
\title{Perform MR 
Perform MR based on the number of SNPs in the dataset}
\usage{
mr_cluster_estimate(dat = dat)
}
\arguments{
\item{dat}{data}
}
\value{
dataframe
}
\description{
Perform MR 
Perform MR based on the number of SNPs in the dataset
}
