% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_plot.R
\name{mr_cluster_forest}
\alias{mr_cluster_forest}
\title{Forest plot}
\usage{
mr_cluster_forest(result = NULL)
}
\arguments{
\item{result}{Results obtained using \code{mr_cluster}}
}
\value{
Plot
}
\description{
Generate forest plot of the MR estimates from each cluster
}
