% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{mr_cluster_heterogeneity}
\alias{mr_cluster_heterogeneity}
\title{Calculate heterogeneity
Calculate heterogeneity summing Q statstics from each cluster}
\usage{
mr_cluster_heterogeneity(cluster = NULL)
}
\arguments{
\item{cluster}{MR estimates from \code{mr_cluster}}
}
\description{
Calculate heterogeneity
Calculate heterogeneity summing Q statstics from each cluster
}
