% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{pam_instruments}
\alias{pam_instruments}
\title{Cluster using partitioning around medoids}
\usage{
pam_instruments(
  bg_dat,
  value_column,
  kmax = min(50, length(unique(bg_dat$SNP))),
  criterion = "asw"
)
}
\arguments{
\item{bg_dat}{<what param does>}

\item{value_column}{<what param does>}

\item{kmax=50}{<what param does>}

\item{criterion="asw"}{<what param does>}
}
\value{

}
\description{
Cluster using partitioning around medoids
}
