% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_plot.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Generate a plot for the simulation result}
\usage{
permutation_plot(simulation = NULL, Q_radial = NULL, Q_mrios = NULL)
}
\arguments{
\item{simulation}{Result from the simulation (permutation test)}

\item{Q_radial}{Q statistics obtained from original radial MR}

\item{Q_mrios}{Q statistics obtained from \code{mr_cluster_heterogeneity}}
}
\value{
Plot
}
\description{
Generate a plot for the result of permutation test
}
