% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{pvclust_instruments}
\alias{pvclust_instruments}
\title{Cluster}
\usage{
pvclust_instruments(
  bg_dat,
  value_column,
  alpha = 0.95,
  method.hclust = "ward.D",
  method.dist = "euclidean",
  nboot = 200
)
}
\arguments{
\item{bg_dat}{<what param does>}

\item{value_column}{<what param does>}

\item{alpha}{<what param does>}

\item{method.hclust}{<what param does>}

\item{method.dist}{<what param does>}

\item{nboot}{<what param does>}
}
\value{

}
\description{
<full description>
}
