% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_background.R
\name{background_ids}
\alias{background_ids}
\title{Generate the list of available GWAS summary statistics}
\usage{
background_ids(
  id_exp = id_exp,
  id_out = id_out,
  type = c("general", "advanced", "eqtl", "metabolite", "protein", "ubm")[1]
)
}
\arguments{
\item{id_exp}{ID for the exposure in IEU GWAS database.}

\item{id_out}{ID for the outome in IEU GWAS database.}

\item{type}{Set a default list of IDs. 
\itemize{
  \item general: Returns list of IDs of GWAS summary statistics generated by many different consortia and generated from UKB.
  \item advanced: In addition to the default list, it returns list of IDs of GWAS summary data on protein, metabolite, and eQTL levels, as well as GWAS summary statistics imported from the EBI database.
  \item eqtl: Returns list of IDs of eQTLGen 2019 results, comprising all cis and some trans regions of gene expression in whole blood.
  \item metabolite: Returns list of IDs of metabolite studies including human blood / immune system / circulating metabolites.
  \item protein: Returns list of IDs of GWAS summary data on protein levels.
  \item ubm: Returns list of IDs of GWAS summary data on brain region volumes. 
}}
}
\value{
List
}
\description{
Returns list of IDs of available GWAS summary statistics. The data is obtained from [IEU GWAS database](https://gwas.mrcieu.ac.uk/).
}
