% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios_IOSMR.R
\name{ios}
\alias{ios}
\title{Calculate index of suspicion}
\usage{
ios(exp = exp_dat, bg = bg_dat)
}
\arguments{
\item{exp}{Instruments for the exposure, obtained using \code{TwoSampleMR::extract_instruments}.}

\item{bg}{Effects for the instruments on a set of variables, used to calculate index of suspicion, obtained using \code{make_background()}.}
}
\value{
Data.frame
}
\description{
Calculate IOS across based on R squared multiple traits and summarising estimates using mean, sd, iqr, median, 95% value, maximum value
}
